Refactor code between vulkan_util and renderer_util
