+++ iris/renderer.h
@@ -80,6 +80,8 @@ void EndFrame(gsl::span<const VkCommandBuffer> secondaryCBs = {}) noexcept;
 tl::expected<absl::FixedArray<VkCommandBuffer>, std::system_error>
 AllocateCommandBuffers(VkCommandBufferLevel level, std::uint32_t count) noexcept;
 
+void AddRenderable(Component::Renderable renderable) noexcept;
+
 /*! \brief Load a file into the rendering system.
  *
  * This is an async load operation, so the only errors returned are if the
